makeParameterContext: use templated isl types
Use explicit API calls instead of operators because the use of aff_set
is confusing (at least to the author of this commit).
The conversions to isl::PwAff are needed because, unlike the Python
bindings, the C++ bindings do not have a class hierarchy.
The conversions could also be avoided by introducing
an isl_aff_nonneg_set and an isl_aff_pos_set in isl.